chart_agent/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
psycopg2-binary
streamlit
requests
python-multipart
python-dotenv
matplotlib
pandas
